clean up warnings from "gcc -Wconversion" (#1344)
* clean up warnings from "gcc -Wconversion"

clean up warnings from "gcc -Wconversion"


* Fix formatting

---------

+++ FreeRTOS-Plus/Demo/FreeRTOS_Plus_TCP_Echo_Qemu_mps2/main_networking.c
@@ -285,7 +285,8 @@ static UBaseType_t uxRand( void )
     /* Utility function to generate a pseudo random number. */
 
     ulNextRand = ( ulMultiplier * ulNextRand ) + ulIncrement;
-    return( ( int ) ( ulNextRand >> 16UL ) & 0x7fffUL );
+
+    return ( ulNextRand >> 16UL ) & 0x7fffUL;
 }
 /*-----------------------------------------------------------*/
